# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OpenDesk Remote API
  version: 0.1.0
  description: REST API skeleton for OpenDesk Remote control plane.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed session bearer token backed by a server-side session hash, or user-bound OpenDesk API token.
    sessionCookie:
      type: apiKey
      in: cookie
      name: opendesk_session
paths:
...
  /api/v1/client-build-configs:
    get:
      summary: List client build configs
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200": { description: Client build configs returned }
        "401": { description: Authentication required }
    post:
      summary: Create client build config
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [build_spec]
              properties:
                name:
                  type: string
                build_spec:
                  type: object
      responses:
        "201": { description: Client build config created }
        "400": { description: Invalid BuildSpec }
        "401": { description: Authentication required }
...
  /api/v1/build-profiles:
    get:
      summary: List build profiles (compatibility alias)
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200": { description: Build profiles returned }
        "401": { description: Authentication required }
    post:
      summary: Create build profile (compatibility alias)
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [build_spec]
              properties:
                name:
                  type: string
                build_spec:
                  type: object
      responses:
        "201": { description: Build profile created }
        "400": { description: Invalid BuildSpec }
        "401": { description: Authentication required }
  /api/v1/build-jobs:
    get:
      summary: List build jobs
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200": { description: Build jobs returned }
        "401": { description: Authentication required }
    post:
      summary: Queue build job
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile_id, platform]
              properties:
                profile_id:
                  type: integer
                platform:
                  type: string
                  enum: [windows_x64, macos_x64, macos_arm64, android_arm64, ios_arm64]
      responses:
        "201": { description: Build job queued }
        "400": { description: Invalid build job }
        "401": { description: Authentication required }
        "404": { description: Build profile not found }
...